requirements.txt
Flask-3.1.0
pandas-2.2.3
sentence-transformers-3.3.1
torch==2.5.1
numpy-2.1.3